break out DescriptorSetProto component
