+++ src/components/Protobuf/DescriptorSetProto/FieldRow/index.js
@@ -1,0 +1,3 @@
+import FieldRow from './FieldRow'
+
+export default FieldRow
