+++ src/components/Protobuf/DescriptorSetProto/InputField/index.js
@@ -1,0 +1,3 @@
+import InputField from './InputField'
+
+export default InputField
