+++ src/components/Protobuf/DescriptorSetProto/SelectOptions/index.js
@@ -1,0 +1,3 @@
+import SelectOptions from './SelectOptions'
+
+export default SelectOptions
