+++ src/components/Protobuf/DescriptorSetProto/constants.js
@@ -1,0 +1,25 @@
+export const protoFields = {
+  string: 'text',
+  bool: 'text',
+  bytes: 'text',
+  double: 'number',
+  float: 'number',
+  enum: 'misc',
+  OipRef: 'text',
+  int32: 'number',
+  int64: 'number',
+  uint32: 'number',
+  uint64: 'number',
+  sint32: 'number',
+  sint64: 'number',
+  fixed32: 'number',
+  fixed64: 'number',
+  sfixed32: 'number',
+  sfixed64: 'number'
+}
+
+export const initialFormRow = {
+  fieldType: 'string',
+  fieldName: '',
+  fieldRule: 'singular'
+}
