+++ src/components/Protobuf/DescriptorSetProto/helpers/shouldUpdate.js
@@ -1,0 +1,7 @@
+const shouldUpdate = (oldProps, newProps) => {
+  const nameDidNotChange = oldProps.state[oldProps.name] === newProps.state[newProps.name]
+  const classesDidNotChange = oldProps.classes === newProps.classes
+  return nameDidNotChange && classesDidNotChange
+}
+
+export default shouldUpdate
